Ajustes ao exibir atributo do tipo número em 'Minhas contribuições' e 'Avaliação de postagens'

+++ implementation/solution/src/main/webapp/modules/admin/controller/markers-controller.js
@@ -2118,7 +2118,7 @@ ront controller of angle won't let enter an invalid URL.
 
                 angular.forEach(result, function (markerAttribute, index) {
                     if (markerAttribute.attribute.type == "NUMBER") {
-                        markerAttribute.value = parseInt(markerAttribute.value);
+                        markerAttribute.value = parseFloat(markerAttribute.value);
                     }
                 });
